refactor(skills): modularize evaluation tooling
