+++ documentation/docs/en/guide/skills.md
@@ -128,6 +128,8 @@ python3 scripts/validate_wow_skills.py
 python3 -m unittest scripts/test_validate_wow_skills.py scripts/test_run_wow_skill_evals.py
 ```
 
+The two entry scripts remain stable, thin CLI façades and support both direct-script and `python -m scripts...` execution. The validator separates core constraints, trace schema, plugin package, eval contracts, and repository orchestration; the runner separates model, I/O, Git, security, state, preparation, evidence, oracles, assertions, verification, and cleanup. New rules belong in the responsibility-specific module under `scripts/wow_skill_validator/` or `scripts/wow_skill_runner/`; entry scripts only parse arguments and orchestrate commands.
+
 The repository-owned validator is self-contained. It validates standard Skill metadata, `openai.yaml`, the explicit plugin list, contained resource paths, shell syntax without executing `--help`, and activation/behavior contracts for all four Skills. [`skills/README.md`](https://github.com/Ahoo-Wang/Wow/blob/main/skills/README.md#validation)
 
 Structural validation cannot prove an eval is executable, automatically activated, or behaviorally correct. `scripts/run_wow_skill_evals.py` freezes the contract and a runtime-only plugin copy without `evals/`, prepares a pinned isolated fixture, and compares a full runner-owned path manifest that records directories and rejects special files. During `prepare`, a protected key domain-signs the RUN v2 descriptor, including the pinned adapter identity and the case, repository, workspace, revision, contract, plugin, and baseline. HMAC evidence independently signs `requestSha256`, which is the SHA-256 of the exact UTF-8 bytes stored in `request.json`, not a canonical-JSON hash. Verification uses only that frozen snapshot and does not reread the live skills package, so an editable `run.json` cannot rebase or redirect a result and asynchronous runs survive later source updates. The standalone review clone receives the exact subject/base object closure in a runner-owned pack before its remote is removed, so detached commits do not depend on advertised refs or external objects. The runner accepts v2 traces only from a pinned adapter with protected HMAC attestation, enforced workspace policy, and commands bound to root cwd plus the resolved executable; activation runs must stop immediately after routing. Authenticated cleanup re-seals its completed lifecycle state; explicit recovery of a damaged marker requires the protected key, uses fixed runner-owned paths plus a caller-supplied source repository, and writes a key-signed idempotency tombstone. Recovery does not cover simultaneous loss of the trust key. Runner-owned oracles independently cover Cart capacity branches, an exact platform contract, and synthetic data interruption/resume, idempotency, checksum, and reconciliation; the interrupted data gate seals an unpredictable, JSON-insignificant byte prefix to reject restart-from-zero rewrites. The hidden Cart test reuses only the workspace-local Gradle cache populated by the signed RED/GREEN commands, forces `--offline` task re-execution, and returns `UNSUPPORTED` rather than `FAIL` when the wrapper, dependency cache, or Java toolchain is unavailable. Missing trust, activation/tool traces, or enforcement also return `UNSUPPORTED`; malformed or tampered evidence returns `ERROR`; assertion failures return `FAIL` and a nonzero CLI exit. See the [runner contract](https://github.com/Ahoo-Wang/Wow/blob/main/skills/README.md#validation) for `prepare`, `verify`, and `cleanup` commands.
