+++ documentation/docs/zh/guide/skills.md
@@ -128,6 +128,8 @@ python3 scripts/validate_wow_skills.py
 python3 -m unittest scripts/test_validate_wow_skills.py scripts/test_run_wow_skill_evals.py
 ```
 
+两个入口脚本保持稳定的薄 CLI façade，并支持脚本与 `python -m scripts...` 两种执行方式。validator 按核心约束、trace schema、插件包、eval contract 和仓库级编排分层；runner 按模型、I/O、Git、安全、状态、准备、证据、oracle、assertion、验证和清理分层。新增规则应进入 `scripts/wow_skill_validator/` 或 `scripts/wow_skill_runner/` 中职责对应的模块，入口只负责参数解析和命令编排。
+
 仓库自带的 validator 不依赖用户目录中的外部脚本；它校验四个 Skill 的标准 metadata、`openai.yaml`、显式插件清单、资源 containment、只做语法检查而不执行 `--help` 的 shell 校验，以及 activation/behavior contract。[`skills/README.md`](https://github.com/Ahoo-Wang/Wow/blob/main/skills/README.md#validation)
 
 结构校验不能证明 eval 可执行、自动激活或行为正确。`scripts/run_wow_skill_evals.py` 会冻结 contract 与去除 `evals/` 的 runtime-only plugin copy，准备固定 revision 的隔离 fixture，并比较会记录目录且拒绝特殊文件的完整路径 manifest。`prepare` 阶段会由受保护 key 对 RUN v2 描述符做 domain-separated 签名，封印固定 adapter identity、case、repository、workspace、revision、contract、plugin 与 baseline；HMAC evidence 再独立签署 `requestSha256`。该值是磁盘上 `request.json` UTF-8 原始文件字节的 SHA-256，而不是 canonical JSON hash。验证只使用这份冻结快照，不回读实时 skills package，因此可编辑的 `run.json` 不能重建 baseline 或重定向结果，异步任务也不受后续源码更新影响。独立 review clone 会先由 runner 以 pack 传输精确 subject/base 的 object closure 再移除 remote，因此 detached commit 不依赖 advertised refs 或外部 object。它只接受固定版本 adapter 通过受保护 HMAC key 签署、workspace policy 已执行、且 command 已绑定根目录 cwd 与解析后 executable 的 v2 trace；activation run 必须在路由后立即停止。受信 cleanup 会重签完成后的生命周期状态；marker 损坏时，显式 recovery 要求保有受保护 key 与操作者提供的 source repository，只清理固定 runner-owned 路径，并写入由 key 签署的 tombstone 保证重试幂等；它不覆盖 trust key 同时丢失的场景。runner-owned oracle 还会独立验证 Cart 容量双分支、精确平台契约，以及 synthetic data 的中断恢复、幂等、checksum 与全量对账；数据中断门禁会封印不可预测、对 JSON 无语义影响的字节前缀，以拒绝从零重写。隐藏 Cart 测试只复用受信 RED/GREEN 命令填充的 workspace-local Gradle cache，强制以 `--offline` 重新执行任务；wrapper、dependency cache 或 Java toolchain 不可用时返回 `UNSUPPORTED` 而非 `FAIL`。缺少 trust、activation/tool trace 或 enforcement 时同样返回 `UNSUPPORTED`，证据畸形或被篡改时返回 `ERROR`，assertion 不满足时返回 `FAIL` 且 CLI 非零退出。`prepare`、`verify`、`cleanup` 的完整命令见 [runner contract](https://github.com/Ahoo-Wang/Wow/blob/main/skills/README.md#validation)。
