+++ scripts/wow_skill_validator/__init__.py
@@ -1,0 +1,16 @@
+"""Modular validation API for repository-owned Wow Skills."""
+
+from . import api, core, evals, package, trace_schema
+from .api import *
+from .core import *
+from .evals import *
+from .package import *
+from .trace_schema import *
+
+__all__ = [
+    *core.__all__,
+    *trace_schema.__all__,
+    *package.__all__,
+    *evals.__all__,
+    *api.__all__,
+]
